add span value in get annotation
